requirements.txt
django
djangorestframework
psycopg2-binary
flake8
#python -m flake8 logistic/

